// src/app/generated/api/model/listFields.ts
/**
 * Listok API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Fields required to create or update a List.
 */
export interface ListFields { 
    /**
     * Name of the List.
     */
    name: string;
    /**
     * Optional comment about the List.
     */
    comment?: string;
    /**
     * Optional image URL for the List.
     */
    img?: string;
}